[~] useRightSidebar refactor
