+++ src/hooks/index.js
@@ -16,6 +16,7 @@ export { default as usePipeline } from 'hooks/usePipeline';
 export { default as useReadme } from 'hooks/useReadme';
 export { default as useRightSidebar } from 'hooks/useRightSidebar.react';
 export { default as useSettings } from 'hooks/useSettings';
+export { default as useStats } from 'hooks/useStats';
 export { default as useStore } from 'hooks/useStore';
 export { default as useTraceData } from 'hooks/useTraceData';
 export { default as useVersions } from 'hooks/useVersions';
