+++ src/hooks/useStats.js
@@ -1,0 +1,12 @@
+import { STATE_SOURCES } from 'const';
+import { useSelector } from 'react-redux';
+
+const useStats = () => {
+  const {
+    dataSource: [cpu, memory],
+  } = useSelector(state => state[STATE_SOURCES.NODE_STATISTICS]);
+
+  return { cpu, memory };
+};
+
+export default useStats;
